Use ESM syntax in src/index.js to allow ESM loaders to work
Using a test setup written in vitest the webflow-api module cannot be imported
because the package.json sets the jsnext:main value to point to src/index.js
which in turn exports code using common js.

Therefore this change uses import/export. The tests still pass and the output to
the dist directory remains the same.